requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
openai==1.46.0
streamlit==1.38.0
python-dotenv==1.0.1
